refactor: ip 찾는 로직 수정, Backup 엔티티 접근제한자 누락 추가, 코드 컨벤션에 맞게 수정

+++ HRBank/src/main/java/org/yebigun/hrbank/domain/backup/dto/BackupDto.java
@@ -16,6 +16,6 @@ public record BackupDto(
     Instant startedAt,
     Instant endedAt,
     BackupStatus status,
-    Integer fileId
+    Long fileId
 ) {
 }
